stacks/ai-ml/04-rag-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
qdrant-client==1.13.3
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.40
minio==7.2.15
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
tiktoken==0.9.0
langchain-text-splitters==0.3.8
python-docx==1.1.2
pypdf==5.4.0
